ai_layer_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Celery & Queue
celery>=5.3.6
redis>=5.0.3

# Database
pymongo>=4.6.2
motor>=3.4.0

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
httpx>=0.27.0
openai>=1.30.0
aioboto3>=13.0.0

# AI/ML
numpy>=1.26.0
langchain>=0.2.0
langchain-openai>=0.1.7
pinecone>=5.0.1
PyPDF2>=3.0.1
cohere>=5.0.0
python-multipart>=0.0.18
google-generativeai>=0.8.0
pdfplumber>=0.11.0
google-genai>=1.0.0

# Table Extraction & Optical/Image
camelot-py[cv]>=0.11.0
pdf2image>=1.17.0
opencv-python>=4.9.0

# Output Builders
python-docx>=1.1.0
openpyxl>=3.1.2
jinja2>=3.1.3
